Adds Error events
It's sometimes beneficial to enable to consumer of the library to decide what happens during an exception event.
This change updates the Connection class to expose an Error event. The Pusher class also now has an Error event.
The Pusher.Error event's implementation will forward connection errors to the consumer of the Pusher instance.