Fixes bug that was causing Subscribed event being fired multiple times
When the Subscribe method is called more then once in a short period of time, multiple successful messages (CHANNEL_SUBSCRIPTION_SUCCEEDED) are received and the Subscribed event was triggered multiple times.

This is not very common, but it can happen especially in cases of connection failure, where the PusherClient tries to reconnect/resubscribe automatically and the client application is also trying to resubscribe to a channel.