feature:(auth) support authorization header with HttpAuthorizer
Using the HTTP authorizer to check access for private channels
requires user access checks. So the HTTP endpoint needs to know
which user is authorized.

HTTP provides "Authorization" header for the task to link an
HTTP request to a user. Unfortunately the HTTP Authorizer has not
supported using such an HTTP header - until now.

The new constructor parameter allows to set an optional
authorization header to the HTTP request.